+++ packages/jest-expo/src/preset/setup.js
@@ -295,3 +295,6 @@ try {
 
 // Ensure the environment globals are installed before the first test runs.
 require('expo/build/winter');
+
+// Ensure the environment variables from dotenv files are loaded before the first test runs
+require('./setup-env').load({ silent: true });